+++ invokeai/app/invocations/model.py
@@ -3,11 +3,13 @@ from typing import List, Optional
 
 from pydantic import BaseModel, ConfigDict, Field
 
+from invokeai.app.shared.fields import FieldDescriptions
+from invokeai.app.shared.models import FreeUConfig
+
 from ...backend.model_management import BaseModelType, ModelType, SubModelType
 from .baseinvocation import (
     BaseInvocation,
     BaseInvocationOutput,
-    FieldDescriptions,
     Input,
     InputField,
     InvocationContext,
@@ -16,22 +18,6 @@ from .baseinvocation import (
     invocation,
     invocation_output,
 )
-
-# TODO: Permanent fix for this
-# from invokeai.app.invocations.shared import FreeUConfig
-
-
-class FreeUConfig(BaseModel):
-    """
-    Configuration for the FreeU hyperparameters.
-    - https://huggingface.co/docs/diffusers/main/en/using-diffusers/freeu
-    - https://github.com/ChenyangSi/FreeU
-    """
-
-    s1: float = Field(ge=-1, le=3, description=FieldDescriptions.freeu_s1)
-    s2: float = Field(ge=-1, le=3, description=FieldDescriptions.freeu_s2)
-    b1: float = Field(ge=-1, le=3, description=FieldDescriptions.freeu_b1)
-    b2: float = Field(ge=-1, le=3, description=FieldDescriptions.freeu_b2)
 
 
 class ModelInfo(BaseModel):
