+++ invokeai/app/shared/__init__.py
@@ -1,0 +1,5 @@
+"""
+This module contains various classes, functions and models which are shared across the app, particularly by invocations.
+
+Lifting these classes, functions and models into this shared module helps to reduce circular imports.
+"""
